# Log correct compaction score for Universal Compaction